+++ sparv/api/util/install.py
@@ -12,9 +12,11 @@ from sparv.api.util import system
 logger = get_logger(__name__)
 
 
-def install_path(source_path: Union[str, Path],
-                 host: Optional[str] = None,
-                 target_path: Optional[Union[str, Path]] = None) -> None:
+def install_path(
+    source_path: Union[str, Path],
+    host: Optional[str],
+    target_path: Union[str, Path]
+) -> None:
     """Transfer a file or the contents of a directory to a target destination, optionally on a different host."""
     system.rsync(source_path, host, target_path)